requirements.txt
absl-py==2.1.0
alabaster==0.7.16
altair==5.4.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
backports.functools-lru-cache==2.0.0
bcrypt==3.2.0
blinker==1.8.2
blis==0.7.11
Brotli==1.1.0
cachetools==5.5.0
cairocffi==1.7.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.44.1
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
constants==0.6.0
contourpy==1.3.0
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.13
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==4.2.11
dnspython==2.7.0
docutils==0.21.2
ecdsa==0.19.1
email_validator==2.2.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.14
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
fpdf==1.7.2
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.11.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
html2image==2.0.7
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface_hub==0.36.2
idna==3.8
imagesize==1.4.1
imbalanced-learn==0.12.4
importlib_metadata==8.5.0
importlib_resources==6.4.5
indic-transliteration==2.3.75
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.7
langcodes==3.5.1
lap==0.5.12
libclang==18.1.1
lxml==6.0.2
mailtrap==2.1.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
ml-dtypes==0.4.1
mpltern==1.0.4
mpmath==1.3.0
mtcnn==0.1.1
murmurhash==1.0.15
mutagen==1.47.0
namex==0.0.8
narwhals==1.35.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
numpydoc==1.8.0
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pangocairocffi==0.7.0
pangocffi==0.13.0
parso==0.8.4
passlib==1.7.4
pdf2image==1.17.0
periodictable==2.0.1
pillow==10.3.0
platformdirs==4.3.3
plotly==6.0.1
preshed==3.0.12
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.26.3
pyparsing==3.1.4
PyPDF2==3.0.1
pypng==0.20220715.0
PyQRCode==1.2.1
pyrolite==0.3.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytube==15.0.0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==8.0
redis==5.2.1
referencing==0.35.1
regex==2024.9.11
reportlab==4.4.5
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rich-toolkit==0.14.8
roman==5.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentence-transformers==5.1.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart_open==7.5.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spotipy==2.24.0
SQLAlchemy==2.0.35
sqlparse==0.5.0
srsly==2.5.2
stack-data==0.6.3
starlette==0.46.2
streamlit==1.38.0
supervision==0.25.1
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tinydb==4.8.2
tokenizers==0.22.2
toml==0.10.2
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.57.6
typer==0.23.2
typer-slim==0.23.2
typing_extensions==4.14.0
tzdata==2024.1
ultralytics==8.3.163
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.35.0
virtualenv==20.26.6
wasabi==1.1.3
watchdog==4.0.2
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.3
websocket-client==1.9.0
websockets==13.0.1
Werkzeug==3.0.4
wrapt==1.16.0
xgboost==2.1.1
yt-dlp==2024.8.6
ytmusicapi==1.8.2
zipp==3.20.2
